Re-evaluation of the logic
Previous version was based on the idea that its always a power of two
It just so happens that the shader I tested this on had power of two all the way.. :+1:
This should make more sense now and be more consistent. Tested with a bunch of shaders and didn't see any logical issues.